controllerでの条件分岐

+++ app/controllers/products_controller.rb
@@ -41,8 +41,12 @@ class ProductsController < ApplicationController
 
   def destroy
     product = Product.find(params[:id])
-    product.destroy
-    redirect_to root_path
+    if product.seller_id == current_user.id && product.destroy
+      product.destroy
+      redirect_to root_path
+    else
+      render :show
+    end
   end
 
   private
